Add a provider which adds Runtime assemblies to metadata reference for VB Framework projects
VB Framework projects needs references to the Core libraries, without which Object and all the essential types will be missing. This reference was added by VisualBasic Project in the legacy world. This change will perform a similar functionality but not at the CPSProject level, the equivalent of the VisualBasic project in the legacy world.

Currently this Provider is added to the Managed VS layer since currently there is no IVT for VB VS layer to ServicesVisualStudio project in Roslyn. Once the change to add IVT to VB VS layer, this provider should move to the VB VS Layer.